sql: fix show create table after altering col type of col with comment
Previously `SHOW CREATE TABLE` use `sub_id` from `system.comments`
as a column id to find a column. However we actually use `PGAttributeNum`
when creating a column comment. So things will break if a column's id
has changed such as when type is altered.

Release note (sql change): Previously, if a column in a table has a comment,
`SHOW CREATE TABLE` would fail after the column type is changed.
Now this is fixed.

+++ pkg/sql/copy.go
@@ -214,7 +214,7 @@ func newCopyMachine(
 			Name:           col.GetName(),
 			Typ:            col.GetType(),
 			TableID:        tableDesc.GetID(),
-			PGAttributeNum: col.GetPGAttributeNum(),
+			PGAttributeNum: uint32(col.GetPGAttributeNum()),
 		}
 	}
 	// If there are no column specifiers and we expect non-visible columns
